perfecting conversion to html

+++ src/harmonic_resonance/chordulator/app.py
@@ -7,16 +7,22 @@ from rich import print, inspect
 from .chordulator import generate_chord_sheet_html
 
 @click.command()
-@click.argument('input_path')
-#  @click.argument('output_file', type=click.File('w'))
-def run(input_path):
-    input_file = Path(input_path)
-    print(input_path)
-    #  inspect(input_file)
-    chord_sheet = input_file.read_text()
-    #  output_path = str(input_file)
-    #  click.echo(output_path)
-    output_file = input_file.with_suffix('.html')
+@click.argument('csml_filename')
+def run(csml_filename):
+    csml_path = Path(csml_filename)
+    
+    # Make sure the csml file exists
+    if not csml_path.exists():
+        print(f"The file {csml_filename} does not exist.")
+        return
+
+    print(csml_filename)
+    chord_sheet = csml_path.read_text()
+    # Compute the output path
+    output_file_path = csml_path.parent / 'html' / csml_path.stem
+    output_file = output_file_path.with_suffix('.html')
+    # Create the html subfolder if it doesn't exist
+    output_file.parent.mkdir(parents=True, exist_ok=True)
     html_output = generate_chord_sheet_html(chord_sheet)
     output_file.write_text(html_output)
     print(output_file)
